Fixed an issue with the Rician noise generator
The function add_rician_noise was normalizing the iid Gausian noise maps to the range [0, 1] before summing with the input image. This is not correct behavior as it means the real component of the noise can only increase the pixel values in the image. I modified the function to correct this behavior.